Added API documentation. Separated algorithm into standalone module 'recursive_match'.
